❗️❗️❗️views.pyをviews/に移行❗️❗️❗️
