+++ docker/srcs/uwsgi-django/pong/static/pong/js/tournament/TournamentCreator.js
@@ -160,6 +160,7 @@ class TournamentCreator
 			UIHelper.putError(error, this.errorMessage);
 		}
 	}
+
 }
 
 export default TournamentCreator;
